Extract functionality to find possible access/egress stops to own class. Put the class behind an interface (RaptorStopFinder) and make it customizable by DependencyInjection. This allows different logic than the default "mode-based concentric circles" to be used to find potential access or egress stops. Sadly, this changes the API a bit to create an instance of SwissRailRaptor.
