+++ src/main/java/ch/sbb/matsim/routing/pt/raptor/SwissRailRaptor.java
@@ -5,22 +5,12 @@
 package ch.sbb.matsim.routing.pt.raptor;
 
 import ch.sbb.matsim.config.SwissRailRaptorConfigGroup;
-import ch.sbb.matsim.config.SwissRailRaptorConfigGroup.IntermodalAccessEgressParameterSet;
 import org.apache.log4j.Logger;
-import org.matsim.api.core.v01.Coord;
 import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.Identifiable;
 import org.matsim.api.core.v01.TransportMode;
-import org.matsim.api.core.v01.network.Link;
 import org.matsim.api.core.v01.population.Leg;
 import org.matsim.api.core.v01.population.Person;
-import org.matsim.api.core.v01.population.PlanElement;
-import org.matsim.api.core.v01.population.Route;
-import org.matsim.core.population.PopulationUtils;
-import org.matsim.core.population.routes.RouteUtils;
-import org.matsim.core.router.RoutingModule;
 import org.matsim.core.utils.geometry.CoordUtils;
-import org.matsim.core.utils.misc.Time;
 import org.matsim.facilities.Facility;
 import org.matsim.pt.router.TransitRouter;
 import org.matsim.pt.transitSchedule.api.TransitStopFacility;
@@ -31,7 +21,6 @@ import java.util.Collection;
 import java.util.Collections;
 import java.util.List;
 import java.util.Map;
-import java.util.stream.Collectors;
 
 /**
  * Provides public transport route search capabilities using an implementation of the
@@ -45,34 +34,31 @@ public class SwissRailRaptor implements TransitRouter {
 
     private final SwissRailRaptorData data;
     private final SwissRailRaptorCore raptor;
-    private final RaptorStaticConfig config;
     private final RaptorParametersForPerson parametersForPerson;
     private final RaptorRouteSelector defaultRouteSelector;
-    private final RaptorIntermodalAccessEgress intermodalAE;
+    private final RaptorStopFinder stopFinder;
     private final String subpopulationAttribute;
     private final ObjectAttributes personAttributes;
-    private final Map<String, RoutingModule> routingModules;
 
     private boolean treeWarningShown = false;
 
     public SwissRailRaptor(final SwissRailRaptorData data, RaptorParametersForPerson parametersForPerson,
-                           RaptorRouteSelector routeSelector, RaptorIntermodalAccessEgress intermodalAE) {
-        this(data, parametersForPerson, routeSelector, intermodalAE, null, null, null);
+                           RaptorRouteSelector routeSelector, RaptorStopFinder stopFinder) {
+        this(data, parametersForPerson, routeSelector, stopFinder, null, null);
         log.info("SwissRailRaptor was initialized without support for subpopulations or intermodal access/egress legs.");
     }
 
     public SwissRailRaptor(final SwissRailRaptorData data, RaptorParametersForPerson parametersForPerson,
-                           RaptorRouteSelector routeSelector, RaptorIntermodalAccessEgress intermodalAE,
-                           String subpopulationAttribute, ObjectAttributes personAttributes, Map<String, RoutingModule> routingModules) {
+                           RaptorRouteSelector routeSelector,
+                           RaptorStopFinder stopFinder,
+                           String subpopulationAttribute, ObjectAttributes personAttributes) {
         this.data = data;
-        this.config = data.config;
         this.raptor = new SwissRailRaptorCore(data);
         this.parametersForPerson = parametersForPerson;
         this.defaultRouteSelector = routeSelector;
-        this.intermodalAE = intermodalAE;
+        this.stopFinder = stopFinder;
         this.subpopulationAttribute = subpopulationAttribute;
         this.personAttributes = personAttributes;
-        this.routingModules = routingModules;
     }
 
     @Override
@@ -197,156 +183,11 @@ public class SwissRailRaptor implements TransitRouter {
     }
 
     private List<InitialStop> findAccessStops(Facility facility, Person person, double departureTime, RaptorParameters parameters) {
-        SwissRailRaptorConfigGroup srrCfg = parameters.getConfig();
-        if (srrCfg.isUseIntermodalAccessEgress()) {
-            return findIntermodalStops(facility, person, departureTime, Direction.Access, parameters);
-        } else {
-            List<TransitStopFacility> stops = findNearbyStops(facility, parameters);
-            List<InitialStop> initialStops = stops.stream().map(stop -> {
-                double beelineDistance = CoordUtils.calcEuclideanDistance(stop.getCoord(), facility.getCoord());
-                double travelTime = Math.ceil(beelineDistance / parameters.getBeelineWalkSpeed());
-                double disutility = travelTime * -parameters.getMarginalUtilityOfTravelTime_utl_s(TransportMode.access_walk);
-                return new InitialStop(stop, disutility, travelTime, beelineDistance, TransportMode.access_walk);
-            }).collect(Collectors.toList());
-            return initialStops;
-        }
+        return this.stopFinder.findStops(facility, person, departureTime, parameters, this.data, RaptorStopFinder.Direction.ACCESS);
     }
 
     private List<InitialStop> findEgressStops(Facility facility, Person person, double departureTime, RaptorParameters parameters) {
-        SwissRailRaptorConfigGroup srrCfg = parameters.getConfig();
-        if (srrCfg.isUseIntermodalAccessEgress()) {
-            return findIntermodalStops(facility, person, departureTime, Direction.Egress, parameters);
-        } else {
-            List<TransitStopFacility> stops = findNearbyStops(facility, parameters);
-            List<InitialStop> initialStops = stops.stream().map(stop -> {
-                double beelineDistance = CoordUtils.calcEuclideanDistance(stop.getCoord(), facility.getCoord());
-                double travelTime = Math.ceil(beelineDistance / parameters.getBeelineWalkSpeed());
-                double disutility = travelTime * -parameters.getMarginalUtilityOfTravelTime_utl_s(TransportMode.egress_walk);
-                return new InitialStop(stop, disutility, travelTime, beelineDistance, TransportMode.egress_walk);
-            }).collect(Collectors.toList());
-            return initialStops;
-        }
-    }
-
-    private enum Direction { Access, Egress }
-
-    private List<InitialStop> findIntermodalStops(Facility facility, Person person, double departureTime, Direction direction, RaptorParameters parameters) {
-        SwissRailRaptorConfigGroup srrCfg = parameters.getConfig();
-        double x = facility.getCoord().getX();
-        double y = facility.getCoord().getY();
-        String personId = person.getId().toString();
-        List<InitialStop> initialStops = new ArrayList<>();
-        for (IntermodalAccessEgressParameterSet paramset : srrCfg.getIntermodalAccessEgressParameterSets()) {
-            double radius = paramset.getRadius();
-            String mode = paramset.getMode();
-            String overrideMode = null;
-            if (mode.equals(TransportMode.walk) || mode.equals(TransportMode.transit_walk)) {
-                overrideMode = direction == Direction.Access ? TransportMode.access_walk : TransportMode.egress_walk;
-            }
-            String linkIdAttribute = paramset.getLinkIdAttribute();
-            String personFilterAttribute = paramset.getPersonFilterAttribute();
-            String personFilterValue = paramset.getPersonFilterValue();
-            String stopFilterAttribute = paramset.getStopFilterAttribute();
-            String stopFilterValue = paramset.getStopFilterValue();
-
-            boolean personMatches = true;
-            if (personFilterAttribute != null) {
-                Object attr = this.personAttributes.getAttribute(personId, personFilterAttribute);
-                String attrValue = attr == null ? null : attr.toString();
-                personMatches = personFilterValue.equals(attrValue);
-            }
-
-            if (personMatches) {
-                Collection<TransitStopFacility> stopFacilities = this.data.stopsQT.getDisk(x, y, radius);
-                for (TransitStopFacility stop : stopFacilities) {
-                    boolean filterMatches = true;
-                    if (stopFilterAttribute != null) {
-                        Object attr = stop.getAttributes().getAttribute(stopFilterAttribute);
-                        String attrValue = attr == null ? null : attr.toString();
-                        filterMatches = stopFilterValue.equals(attrValue);
-                    }
-                    if (filterMatches) {
-                        Facility stopFacility = stop;
-                        if (linkIdAttribute != null) {
-                            Object attr = stop.getAttributes().getAttribute(linkIdAttribute);
-                            if (attr != null) {
-                                stopFacility = new ChangedLinkFacility(stop, Id.create(attr.toString(), Link.class));
-                            }
-                        }
-
-                        List<? extends PlanElement> routeParts;
-                        if (direction == Direction.Access) {
-                            RoutingModule module = this.routingModules.get(mode);
-                            routeParts = module.calcRoute(facility, stopFacility, departureTime, person);
-                        } else { // it's Egress
-                            // We don't know the departure time for the egress trip, so just use the original departureTime,
-                            // although it is wrong and might result in a wrong traveltime and thus wrong route.
-                            RoutingModule module = this.routingModules.get(mode);
-                            routeParts = module.calcRoute(stopFacility, facility, departureTime, person);
-                            // clear the (wrong) departureTime so users don't get confused
-                            for (PlanElement pe : routeParts) {
-                                if (pe instanceof Leg) {
-                                    ((Leg) pe).setDepartureTime(Time.getUndefinedTime());
-                                }
-                            }
-                        }
-                        if (overrideMode != null) {
-                            for (PlanElement pe : routeParts) {
-                                if (pe instanceof Leg) {
-                                    ((Leg) pe).setMode(overrideMode);
-                                }
-                            }
-                        }
-                        if (stopFacility != stop) {
-                            if (direction == Direction.Access) {
-                                Leg transferLeg = PopulationUtils.createLeg(TransportMode.transit_walk);
-                                Route transferRoute = RouteUtils.createGenericRouteImpl(stopFacility.getLinkId(), stop.getLinkId());
-                                transferRoute.setTravelTime(0);
-                                transferRoute.setDistance(0);
-                                transferLeg.setRoute(transferRoute);
-                                transferLeg.setTravelTime(0);
-
-                                List<PlanElement> tmp = new ArrayList<>(routeParts.size() + 1);
-                                tmp.addAll(routeParts);
-                                tmp.add(transferLeg);
-                                routeParts = tmp;
-                            } else {
-                                Leg transferLeg = PopulationUtils.createLeg(TransportMode.transit_walk);
-                                Route transferRoute = RouteUtils.createGenericRouteImpl(stop.getLinkId(), stopFacility.getLinkId());
-                                transferRoute.setTravelTime(0);
-                                transferRoute.setDistance(0);
-                                transferLeg.setRoute(transferRoute);
-                                transferLeg.setTravelTime(0);
-
-                                List<PlanElement> tmp = new ArrayList<>(routeParts.size() + 1);
-                                tmp.add(transferLeg);
-                                tmp.addAll(routeParts);
-                                routeParts = tmp;
-                            }
-                        }
-                        RaptorIntermodalAccessEgress.RIntermodalAccessEgress accessEgress = this.intermodalAE.calcIntermodalAccessEgress(routeParts, parameters);
-                        InitialStop iStop = new InitialStop(stop, accessEgress.disutility, accessEgress.travelTime, accessEgress.routeParts);
-                        initialStops.add(iStop);
-                    }
-                }
-            }
-        }
-        return initialStops;
-    }
-
-    private List<TransitStopFacility> findNearbyStops(Facility facility, RaptorParameters parameters) {
-        double x = facility.getCoord().getX();
-        double y = facility.getCoord().getY();
-        Collection<TransitStopFacility> stopFacilities = this.data.stopsQT.getDisk(x, y, parameters.getSearchRadius());
-        if (stopFacilities.size() < 2) {
-            TransitStopFacility  nearestStop = this.data.stopsQT.getClosest(x, y);
-            double nearestDistance = CoordUtils.calcEuclideanDistance(facility.getCoord(), nearestStop.getCoord());
-            stopFacilities = this.data.stopsQT.getDisk(x, y, nearestDistance + parameters.getExtensionRadius());
-        }
-        if (stopFacilities instanceof List) {
-            return (List<TransitStopFacility>) stopFacilities;
-        }
-        return new ArrayList<>(stopFacilities);
+        return this.stopFinder.findStops(facility, person, departureTime, parameters, this.data, RaptorStopFinder.Direction.EGRESS);
     }
 
     private RaptorRoute createDirectWalk(Facility fromFacility, Facility toFacility, double departureTime, Person person, RaptorParameters parameters) {
@@ -358,37 +199,6 @@ public class SwissRailRaptor implements TransitRouter {
         RaptorRoute route = new RaptorRoute(fromFacility, toFacility, walkCost);
         route.addNonPt(null, null, departureTime, walkTime, beelineDistance, TransportMode.transit_walk);
         return route;
-    }
-
-    private static class ChangedLinkFacility implements Facility, Identifiable<TransitStopFacility> {
-
-        private final TransitStopFacility delegate;
-        private final Id<Link> linkId;
-
-        ChangedLinkFacility(final TransitStopFacility delegate, final Id<Link> linkId) {
-            this.delegate = delegate;
-            this.linkId = linkId;
-        }
-
-        @Override
-        public Id<Link> getLinkId() {
-            return this.linkId;
-        }
-
-        @Override
-        public Coord getCoord() {
-            return this.delegate.getCoord();
-        }
-
-        @Override
-        public Map<String, Object> getCustomAttributes() {
-            return this.delegate.getCustomAttributes();
-        }
-
-        @Override
-        public Id<TransitStopFacility> getId() {
-            return this.delegate.getId();
-        }
     }
 
 }
