+++ src/main/java/ch/sbb/matsim/routing/pt/raptor/SwissRailRaptorFactory.java
@@ -4,23 +4,17 @@
 
 package ch.sbb.matsim.routing.pt.raptor;
 
-import ch.sbb.matsim.config.SwissRailRaptorConfigGroup;
-import ch.sbb.matsim.config.SwissRailRaptorConfigGroup.IntermodalAccessEgressParameterSet;
 import org.matsim.api.core.v01.network.Network;
 import org.matsim.api.core.v01.population.Population;
 import org.matsim.core.api.experimental.events.EventsManager;
 import org.matsim.core.config.Config;
-import org.matsim.core.config.ConfigUtils;
 import org.matsim.core.config.groups.PlansConfigGroup;
-import org.matsim.core.router.RoutingModule;
 import org.matsim.pt.router.TransitScheduleChangedEventHandler;
 import org.matsim.pt.transitSchedule.api.TransitSchedule;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
 import javax.inject.Singleton;
-import java.util.HashMap;
-import java.util.Map;
 
 /**
  * @author mrieser / SBB
@@ -33,54 +27,35 @@ public class SwissRailRaptorFactory implements Provider<SwissRailRaptor> {
     private final RaptorStaticConfig raptorConfig;
     private final RaptorParametersForPerson raptorParametersForPerson;
     private final RaptorRouteSelector routeSelector;
-    private final RaptorIntermodalAccessEgress intermodalAE;
-
+    private final Provider<RaptorStopFinder> stopFinderProvider;
     private final Network network;
     private final PlansConfigGroup plansConfigGroup;
     private final Population population;
-    private final Map<String, Provider<RoutingModule>> routingModuleProviders;
 
     @Inject
     public SwissRailRaptorFactory(final TransitSchedule schedule, final Config config, final Network network,
                                   RaptorParametersForPerson raptorParametersForPerson, RaptorRouteSelector routeSelector,
-                                  RaptorIntermodalAccessEgress intermodalAE, PlansConfigGroup plansConfigGroup, Population population,
-                                  Map<String, Provider<RoutingModule>> routingModules, final EventsManager events) {
+                                  Provider<RaptorStopFinder> stopFinderProvider, PlansConfigGroup plansConfigGroup, Population population,
+                                  final EventsManager events) {
         this.schedule = schedule;
         this.raptorConfig = RaptorUtils.createStaticConfig(config);
         this.network = network;
         this.raptorParametersForPerson = raptorParametersForPerson;
         this.routeSelector = routeSelector;
-        this.intermodalAE = intermodalAE;
+        this.stopFinderProvider = stopFinderProvider;
         this.plansConfigGroup = plansConfigGroup;
         this.population = population;
 
-        SwissRailRaptorConfigGroup srrConfig = ConfigUtils.addOrGetModule(config, SwissRailRaptorConfigGroup.class);
-        this.routingModuleProviders = new HashMap<>();
-        if (srrConfig.isUseIntermodalAccessEgress()) {
-            for (IntermodalAccessEgressParameterSet params : srrConfig.getIntermodalAccessEgressParameterSets()) {
-                String mode = params.getMode();
-                this.routingModuleProviders.put(mode, routingModules.get(mode));
-            }
-        }
-        
         if (events != null) {
-            events.addHandler((TransitScheduleChangedEventHandler) event -> {
-                this.data = null;
-            });
+            events.addHandler((TransitScheduleChangedEventHandler) event -> this.data = null);
         }
     }
 
     @Override
     public SwissRailRaptor get() {
         SwissRailRaptorData data = getData();
-        Map<String, RoutingModule> neededRoutingModules = new HashMap<>();
-        for (Map.Entry<String, Provider<RoutingModule>> e : this.routingModuleProviders.entrySet()) {
-            String mode = e.getKey();
-            RoutingModule module = e.getValue().get();
-            neededRoutingModules.put(mode, module);
-        }
-        return new SwissRailRaptor(data, this.raptorParametersForPerson, this.routeSelector, this.intermodalAE,
-                this.plansConfigGroup.getSubpopulationAttributeName(), this.population.getPersonAttributes(), neededRoutingModules);
+        return new SwissRailRaptor(data, this.raptorParametersForPerson, this.routeSelector, this.stopFinderProvider.get(),
+                this.plansConfigGroup.getSubpopulationAttributeName(), this.population.getPersonAttributes());
     }
 
     private SwissRailRaptorData getData() {
